+++ src/integration-tests/Helpers/Scripts.hs
@@ -5,6 +5,7 @@ import Hasql.Connection.Setting.Connection qualified as Connection.Setting.Conne
 import Hasql.Connection.Setting.Connection.Param qualified as Connection.Setting.Connection.Param
 import Hasql.Pool qualified as Pool
 import Hasql.Pool.Config qualified as Config
+import Hasql.Session qualified as Session
 import System.Random.Stateful qualified as Random
 import TextBuilder qualified
 import Prelude
@@ -45,6 +46,38 @@ onAutotaggedPool poolSize acqTimeout maxLifetime maxIdletime (host, port) cont =
   -- Generate app name
   appName <- generateName "hasql-pool-test-"
   onTaggedPool poolSize acqTimeout maxLifetime maxIdletime appName (host, port) (cont appName)
+
+onTaggedPoolWithInitSession :: Int -> DiffTime -> DiffTime -> DiffTime -> Session.Session () -> Text -> ScopeParams -> (Pool.Pool -> IO ()) -> IO ()
+onTaggedPoolWithInitSession poolSize acqTimeout maxLifetime maxIdletime initSession appName (host, port) =
+  bracket
+    ( Pool.acquire
+        ( Config.settings
+            [ Config.size poolSize,
+              Config.acquisitionTimeout acqTimeout,
+              Config.agingTimeout maxLifetime,
+              Config.idlenessTimeout maxIdletime,
+              Config.initSession initSession,
+              Config.staticConnectionSettings
+                [ Connection.Setting.connection
+                    ( Connection.Setting.Connection.params
+                        [ Connection.Setting.Connection.Param.host host,
+                          Connection.Setting.Connection.Param.port (fromIntegral port),
+                          Connection.Setting.Connection.Param.user "postgres",
+                          Connection.Setting.Connection.Param.password "",
+                          Connection.Setting.Connection.Param.dbname "postgres",
+                          Connection.Setting.Connection.Param.other "application_name" appName
+                        ]
+                    )
+                ]
+            ]
+        )
+    )
+    Pool.release
+
+onAutotaggedPoolWithInitSession :: Int -> DiffTime -> DiffTime -> DiffTime -> Session.Session () -> ScopeParams -> (Text -> Pool.Pool -> IO ()) -> IO ()
+onAutotaggedPoolWithInitSession poolSize acqTimeout maxLifetime maxIdletime initSession (host, port) cont = do
+  appName <- generateName "hasql-pool-test-"
+  onTaggedPoolWithInitSession poolSize acqTimeout maxLifetime maxIdletime initSession appName (host, port) (cont appName)
 
 onDefaultTaggedPool :: ScopeParams -> (Text -> Pool.Pool -> IO ()) -> IO ()
 onDefaultTaggedPool =