+++ src/library/exposed/Hasql/Pool.hs
@@ -86,7 +86,7 @@ acquire config = do
     join . atomically $ do
       entries <- flushTQueue connectionQueue
       let (agedEntries, unagedEntries) = partition (entryIsAged agingTimeoutNanos now) entries
-          (idleEntries, liveEntries) = partition (entryIsIdle agingTimeoutNanos now) unagedEntries
+          (idleEntries, liveEntries) = partition (entryIsIdle maxIdletimeNanos now) unagedEntries
       traverse_ (writeTQueue connectionQueue) liveEntries
       return $ do
         forM_ agedEntries $ \entry -> do
@@ -177,6 +177,7 @@ use Pool {..} sess = do
           Session.run poolInitSession connection >>= \case
             Left err -> do
               Connection.release connection
+              atomically $ modifyTVar' poolCapacity succ
               ErrorsDestruction.reset
                 ( \details -> do
                     poolObserver (ConnectionObservation id (TerminatedConnectionStatus (NetworkErrorConnectionTerminationReason (fmap (Text.decodeUtf8With Text.lenientDecode) details))))